layers: Create a Stateless Util source file
