+++ BUILD.gn
@@ -239,6 +239,7 @@ stateless_validation_sources = [
   "layers/stateless/sl_ray_tracing.cpp",
   "layers/stateless/sl_render_pass.cpp",
   "layers/stateless/sl_synchronization.cpp",
+  "layers/stateless/sl_utils.cpp",
   "layers/stateless/sl_wsi.cpp",
   "layers/stateless/stateless_validation.h",
 ]
